move relevant tests to OpenHABCore
