+++ OpenHABCore/Sources/OpenHABCore/Model/OpenHABItem.swift
@@ -13,6 +13,7 @@ import CoreLocation
 import Fuzi
 import os.log
 import UIKit
+
 public final class OpenHABItem: NSObject, CommItem {
     public enum ItemType: String {
         case color = "Color"
