+++ OpenHABCore/Tests/OpenHABCoreTests/JSONData.swift
