+++ OpenHABCore/Tests/OpenHABCoreTests/MockURLProtocol.swift
