--- OpenHABCore/Tests/OpenHABCoreTests/OpenHABCoreTests.swift
@@ -1,22 +1,0 @@
-// Copyright (c) 2010-2022 Contributors to the openHAB project
-//
-// See the NOTICE file(s) distributed with this work for additional
-// information.
-//
-// This program and the accompanying materials are made available under the
-// terms of the Eclipse Public License 2.0 which is available at
-// http://www.eclipse.org/legal/epl-2.0
-//
-// SPDX-License-Identifier: EPL-2.0
-
-@testable import OpenHABCore
-import XCTest
-
-final class OpenHABCoreTests: XCTestCase {
-    func testExample() throws {
-        // This is an example of a functional test case.
-        // Use XCTAssert and related functions to verify your tests produce the correct
-        // results.
-        XCTAssertEqual(OpenHABCore().text, "Hello, World!")
-    }
-}
