+++ OpenHABCore/Tests/OpenHABCoreTests/ParseAsTests.swift
@@ -9,12 +9,11 @@
 //
 // SPDX-License-Identifier: EPL-2.0
 
-import XCTest
-
-@testable import openHAB
 @testable import OpenHABCore
 
-class ParseAsTest: XCTestCase {
+import XCTest
+
+final class ParseAsTest: XCTestCase {
     override func setUpWithError() throws {
         // Put setup code here. This method is called before the invocation of each test method in the class.
     }
