+++ OpenHABCore/Tests/OpenHABCoreTests/XMLData.swift
